cleans up integration tests
