+++ packages/shared-modules/src/modules/core/lib/logger/index.ts
@@ -20,18 +20,18 @@ const resolveLogger = () => {
     return new SentryLogger(process.env.NF_SENTRY_DSN);
   }
 
-  /*   if (
+  if (
     process.env.NODE_ENV === "production" &&
     process.env.TYPE_OF_DEPLOYMENT !== "commit" &&
     process.env.NF_CYPRESS_RUN === "0"
   ) {
     // tslint:disable-next-line
-    console.info("Error logging is disabled"); */
+    console.info("Error logging is disabled");
 
-  return noopLogger;
-  // }
+    return noopLogger;
+  }
 
-  // return new DevConsoleLogger();
+  return new DevConsoleLogger();
 };
 
 @injectable()
