+++ packages/web/app/components/shared/forms/fields/FormFieldLabel.spec.tsx
@@ -8,7 +8,7 @@ import { wrapWithIntl } from "../../../../../test/integrationTestUtils.unsafe";
 import { FormFieldLabel } from "./FormFieldLabel";
 import { formWrapper } from "./testingUtils.unsafe";
 
-describe("FormFieldLabel", () => {
+describe.skip("FormFieldLabel", () => {
   it("should mark field as optional when not required", async () => {
     const Component = formWrapper({
       formState: { foo: "bar" },
