+++ packages/web/app/modules/wallet-selector/forms/sagas.ts
@@ -41,7 +41,6 @@ export function* registerForm(
     userEmail?: string;
   },
 ): Generator<any, void, any> {
-  
   yield put(
     actions.walletSelector.setWalletRegisterData({
       ...baseUiData,
