+++ packages/web/app/modules/wallet-selector/sagas.ts
@@ -1,18 +1,11 @@
-import {
-  fork,
-  neuCall,
-  neuTakeEvery,
-  neuTakeLatestUntil,
-  put,
-  select,
-} from "@neufund/sagas";
+import { fork, neuCall, neuTakeEvery, neuTakeLatestUntil, put, select } from "@neufund/sagas";
 
 import { TAppGlobalState } from "../../store";
 import { actions } from "../actions";
 import { handleSignInUser } from "../auth/user/sagas";
 import { loadPreviousWallet } from "../web3/sagas";
 import { EWalletType } from "../web3/types";
-import { ledgerRegister } from './ledgerConnectAndSign';
+import { ledgerRegister } from "./ledgerConnectAndSign";
 import { walletSelectorInitialState } from "./reducer";
 import { selectUrlUserType } from "./selectors";
 import { EFlowType } from "./types";
@@ -35,7 +28,6 @@ export function* walletSelectorConnect(): Generator<any, any, any> {
 export function* walletSelectorReset(): Generator<any, any, any> {
   yield neuCall(loadPreviousWallet);
 }
-
 
 export function* resetWalletSelectorState(): Generator<any, void, any> {
   yield put(actions.walletSelector.reset());
