+++ packages/web/app/test-e2e/investor/portfolio/TokenTransfer.spec.e2e.io.ts
@@ -1,10 +1,10 @@
+import { DEFAULT_PASSWORD } from './../../utils/constants';
 import { getTokenBalance } from "../../utils/ethRpcUtils";
 import { confirmAccessModal, etoFixtureAddressByName, etoFixtureByName } from "../../utils/index";
 import { goToPortfolio } from "../../utils/navigation";
 import { tid } from "../../utils/selectors";
 import {
   createAndLoginNewUser,
-  DEFAULT_PASSWORD,
   loginFixtureAccount,
 } from "../../utils/userHelpers";
 import { assertWithdrawButtonIsDisabled, typeWithdrawForm } from "../wallet/utils";
