+++ packages/web/app/test-e2e/issuer/eto/EtoRegistrationUtils.ts
@@ -1,6 +1,7 @@
 import { toCamelCase, toSnakeCase, withParams } from "@neufund/shared";
 import { isFunction } from "lodash/fp";
 
+import { createUser } from "./../../utils/createUser";
 import { etoRegisterRoutes } from "../../../components/eto/registration/routes";
 import { TPartialCompanyEtoData } from "../../../lib/api/eto/EtoApi.interfaces.unsafe";
 import { confirmAccessModal } from "../../utils";
@@ -9,7 +10,7 @@ import { cyPromise } from "../../utils/cyPromise";
 import { checkForm, fillForm, TFormFixture, TFormFixtureExpectedValues } from "../../utils/forms";
 import { goToIssuerDashboard } from "../../utils/navigation";
 import { tid } from "../../utils/selectors";
-import { createUser, makeAuthenticatedCall } from "../../utils/userHelpers";
+import { makeAuthenticatedCall } from "../../utils/userHelpers";
 import {
   aboutFormRequired,
   aboutFormSubmit,
