+++ packages/web/app/test-e2e/other/recovery/Backup.spec.e2e.io.ts
@@ -1,10 +1,10 @@
+import { DEFAULT_PASSWORD } from "./../../utils/constants";
 import { LIGHT_WALLET_PRIVATE_DATA_CACHE_TIME } from "../../../config/constants";
 import { backupLightWalletSeed } from "../../utils/backupLightWalletSeed";
 import {
   accountFixturePrivateKey,
   assertLockedAccessModal,
   confirmAccessModal,
-  DEFAULT_PASSWORD,
   generateRandomEmailAddress,
   goToProfile,
   loginFixtureAccount,
