+++ packages/web/app/test-e2e/utils/constants.ts
@@ -1,0 +1,8 @@
+export const DEFAULT_PASSWORD = "strongpassword";
+export const DEFAULT_HD_PATH = "m/44'/60'/0'";
+
+export const VAULT_API_ROOT = "/api/wallet";
+export const WALLET_STORAGE_KEY = "NF_WALLET_METADATA";
+export const JWT_KEY = "NF_JWT";
+export const NF_USER_KEY = "NF_USER";
+export type TKycType = "business" | "individual";
