+++ packages/web/health-check-sha.json
@@ -1,1 +1,1 @@
-{ "artifactsSha": "99b098c43", "commitSha": "f3f1f65" }
+{ "artifactsSha": "5e4b50c69", "commitSha": "c4b8145" }
